gpt-5-5: split ProjectsSection test monolith into focused suites (#260)
Decomposed 917-line ProjectsSection.test.tsx into rendering (19), scroll
(23), and cards (24) suites plus shared fixtures. PROJECT_CARD_WIDTH
assertion imports projectsGeometry per #258 layout contract (layout.ts
holds NAV_HEIGHT only). No component changes.

Files: ProjectsSection.{rendering,scroll,cards}.test.tsx,
ProjectsSection.test.shared.ts; removed ProjectsSection.test.tsx.

Notes: npm run test has 2 pre-existing architecture.test.ts failures on
main (resume-audit timelineEntries import); all 66 projects tests pass.
